xplist: Plug memory leak by cleaning up libxml2's parser after use
This is actually considered bad practice. However, it appears this
memory leak is otherwise not possible to fix due to a design flaw
in how libxml2 handles the lifecycle of it's XML parser.
We'll let the community test this in production now and decide.
In our tests this change had no drawbacks except fixing the last known
memory leak in libplist.