Fix for LibTool compalation and Python 3 libintl
On Python 3.9 `libpython` no longer is linkable as a static library
due to the fact that `libpython` now depends on `libintl`.  This would
mean we would have to import `libintl` to create a fully linked .la

It is better to be explicit that we are building a .so (really a .dylib
but autotools uses linux file conventions) that doesn't have to be fully
resolved.